requirements.txt
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
ftfy==5.5.1
idna==2.10
lxml==4.6.2
numpy==1.19.5
pandas==0.25.3
Pillow==8.1.0
prettytable==2.0.0
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.5
quantipy3==0.2.3
requests==2.25.1
savReaderWriter==3.4.2
scipy==1.6.0
six==1.15.0
urllib3==1.26.2
watchdog==1.0.2
wcwidth==0.2.5
XlsxWriter==1.3.7
xmltodict==0.12.0
